rough draft of cleaning up components
