+++ src/styles/main.scss
@@ -13,7 +13,7 @@
 @import './base/_mixins.scss';
 @import './../components/location-finder/Hero/Hero.scss';
 @import './../components/location-finder/SearchForm/SearchForm.scss';
-@import './../components/location-finder/SearchForm/SearchAutoSuggest.scss';
+@import './../components/location-finder/SearchAutoSuggest/SearchAutoSuggest.scss';
 @import './../components/location-finder/SearchHeader/SearchHeader.scss';
 @import './../components/location-finder/Locations/Locations.scss';
 @import './../components/location-finder/Location/Location.scss';
@@ -36,5 +36,3 @@
     padding: 0 1.5rem;
   }
 }
-
-
